requirements.txt
Flask==2.x.x
mysql-connector-python==8.x.x